feat: use es2019 for lib and target
We are currently using node 12 in all of our node TypeScript projects,
and we are beginning to use babel to transpile front-end projects. So,
to take advantage of new features in TypeScript/JavaScript we need to
increase our tsconfig lib and target settings. Since node 12 supports
all of es2019, this seems a good version to target.

Also, as we upgrade to the current version of TypeScript (3.9.2) some of
the functionality that was allowed in previous versions of TypeScript
when targeting "es5" now require tslib. For example, array spread notation
now requires tslib to transpile to es5.